requirements.txt
plotly
streamlit
pandas
numpy
psycopg2-binary
